Bumps lagom version to 1.3.5 and provides lagom scaladsl
