+++ lagom-service-locator-javadsl-consul/src/main/resources/reference.conf
@@ -1,0 +1,10 @@
+#############################################################
+# Lagom Service Locator for ZooKeeper Reference Config File #
+#############################################################
+
+# This is the reference config file that contains all the default settings.
+# Make your edits/overrides in your application.conf.
+
+# Enables the `ConsulServiceLocatorModule` to register the `ZooKeeperServiceLocator`.
+# The `ZooKeeperServiceLocator` implements Lagom's ServiceLocator
+play.modules.enabled += "com.lightbend.lagom.javadsl.discovery.consul.ConsulServiceLocatorModule"
