+++ lagom-service-locator-javadsl-consul/src/test/resources/application.conf
