+++ lagom-service-locator-scaladsl-consul/src/test/resources/application.conf
@@ -1,0 +1,10 @@
+lagom {
+  discovery {
+    consul {
+      agent-hostname = "127.0.0.1"   # hostname or IP-address for the Consul agent
+      agent-port     = 8500          # port for the Consul agent
+      uri-scheme     = "http"        # for example: http or https
+      routing-policy = "round-robin" # valid routing policies: first, random, round-robin
+    }
+  }
+}
