--- src/main/resources/reference.conf
@@ -1,10 +1,0 @@
-#############################################################
-# Lagom Service Locator for ZooKeeper Reference Config File #
-#############################################################
-
-# This is the reference config file that contains all the default settings.
-# Make your edits/overrides in your application.conf.
-
-# Enables the `ConsulServiceLocatorModule` to register the `ZooKeeperServiceLocator`.
-# The `ZooKeeperServiceLocator` implements Lagom's ServiceLocator
-play.modules.enabled += "com.lightbend.lagom.discovery.consul.ConsulServiceLocatorModule"
